Refactor Allonsh into modular architecture and Rename allonsh.js to Allonsh.js
